+++ datacontract/lint/urls.py
@@ -21,7 +21,7 @@ def fetch_resource(url: str, config: Config | None = None):
             type="lint",
             name=f"Reading data contract from {url}",
             reason=f"Cannot read resource from URL {url}. Response status is {response.status_code}",
-            engine="datacontract",
+            engine="datacontract-cli",
             result=ResultEnum.error,
         )
 
@@ -42,7 +42,7 @@ def _set_api_key(headers, url, config: Config):
             type="lint",
             name=f"Reading data contract from {url}",
             reason="Error: Entropy Data API key is not set. Set env variable ENTROPY_DATA_API_KEY.",
-            engine="datacontract",
+            engine="datacontract-cli",
             result=ResultEnum.error,
         )
     headers["x-api-key"] = api_key